# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Ludus Engine API",
    "description" : "Backend engine for 2D games. Content is authored in the editor and served to game clients over the public routes, which are cacheable and validated by ETag.",
    "license" : {
      "name" : "AGPL-3.0-or-later",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version" : "${project.version}"
  },
  "paths" : {
    "/api/v1/admin/waves/{id}" : {
      "get" : {
        "tags" : [ "Wave authoring" ],
        "summary" : "Fetch a wave's document, draft or not",
        "description" : "Returns the stored bytes exactly as they were received.",
        "operationId" : "getWaveDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Wave authoring" ],
        "summary" : "Replace a wave's document",
        "description" : "The id in the URL and the id in the document must agree.",
        "operationId" : "replaceWave",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Wave authoring" ],
        "summary" : "Delete a wave",
        "operationId" : "deleteWave",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/admin/wave-levels/{id}" : {
      "get" : {
        "tags" : [ "Wave levels" ],
        "summary" : "One level",
        "operationId" : "getWaveLevel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Wave levels" ],
        "summary" : "Rename or resequence a level",
        "description" : "The wave list is replaced wholesale, in the order given.",
        "operationId" : "updateWaveLevel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Wave levels" ],
        "summary" : "Delete a level",
        "description" : "The waves themselves are untouched; only their membership of this level goes. Deleting the active level leaves the project with none active.",
        "operationId" : "deleteWaveLevel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/wave-levels" : {
      "get" : {
        "tags" : [ "Wave levels" ],
        "summary" : "Every level in the project",
        "operationId" : "listWaveLevels",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Summary"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Wave levels" ],
        "summary" : "Create a level",
        "description" : "The waves are listed in play order. Creating a level never activates it; activation is a separate call.",
        "operationId" : "createWaveLevel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/wave-levels/{id}/activate" : {
      "post" : {
        "tags" : [ "Wave levels" ],
        "summary" : "Make this the level players receive",
        "description" : "Replaces whatever was active. A project has at most one active level, which the database guarantees rather than this endpoint.",
        "operationId" : "activateWaveLevel",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/waves/{id}" : {
      "get" : {
        "tags" : [ "Public content" ],
        "summary" : "One published wave's indexed fields",
        "operationId" : "getPublishedWave",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "If-None-Match",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Summary" : {
        "type" : "object",
        "description" : "A wave's indexed fields. The document itself is fetched separately.",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "order" : {
            "type" : "integer",
            "format" : "int32"
          },
          "schemaVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "published" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Request" : {
        "type" : "object",
        "description" : "A level's name and the waves it sequences, in play order.",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "waves" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    }
  }
}